remove download from store
