+++ src/constants/actionTypes.js
@@ -26,8 +26,6 @@ export const UPDATE_FIELD_REPORT = 'UPDATE_FIELD_REPORT';
 export const UPDATE_FIELD_REPORT_PATIENT = 'UPDATE_FIELD_REPORT_PATIENT';
 export const UPDATE_FIELD_REPORT_THERAPIST = 'UPDATE_FIELD_REPORT_THERAPIST';
 
-export const REPORT_PDF_CREATED = 'REPORT_PDF_CREATED';
-export const REPORT_DOWNLOADED = 'REPORT_DOWNLOADED';
 export const REPORT_UPLOADED = 'REPORT_UPLOADED';
 export const EDITOR_SIDENAV_LOADED = 'EDITOR_SIDENAV_LOADED';
 export const EDITOR_SIDENAV_UNLOADED = 'EDITOR_SIDENAV_UNLOADED';
