+++ src/reducers/editorSideNav.js
@@ -1,6 +1,4 @@
 import {
-    REPORT_PDF_CREATED,
-    REPORT_DOWNLOADED,
     EDITOR_SIDENAV_LOADED,
     EDITOR_SIDENAV_UNLOADED
 } from "../constants/actionTypes";
@@ -13,15 +11,6 @@ export default (state = {}, action) => {
             };
         case EDITOR_SIDENAV_UNLOADED:
             return {};
-        case REPORT_PDF_CREATED:
-            return {
-                blob: action.error ? null : action.payload,
-                errors: action.error ? {"pdf creation":["an error occured"]} : null
-            };
-        case REPORT_DOWNLOADED:
-            return {
-                download: action.payload
-            };
         default:
             return state;
     }
