+++ src/reducers/home.js
@@ -1,8 +1,7 @@
 import {
     HOME_PAGE_LOADED,
     HOME_PAGE_UNLOADED,
-    DELETE_REPORT,
-    REPORT_PDF_CREATED
+    DELETE_REPORT
 } from '../constants/actionTypes';
 
 export default (state = {}, action) => {
@@ -20,12 +19,6 @@ export default (state = {}, action) => {
 				reports: state.reports.filter(r => r.id !== action.reportId),
                 errors: action.error ? action.payload.errors : null
 			};
-        case REPORT_PDF_CREATED:
-            return {
-                ...state,
-                blob: action.error ? null : action.payload,
-                errors: action.error ? {"pdf creation":["an error occured"]} : null
-            };
         default:
             return state;
     }
