+++ src/utils/downloadPdf.js
@@ -1,0 +1,36 @@
+export default function(res) {
+    let filename = "";
+    const disposition = res.header['content-disposition'];
+    if (disposition && disposition.indexOf('attachment') !== -1) {
+        const filenameRegex = /filename[^;=\n]*=((['"]).*?\2|[^;\n]*)/;
+        const matches = filenameRegex.exec(disposition);
+        if (matches != null && matches[1]) filename = matches[1].replace(/['"]/g, '');
+    }
+    const blob = new Blob([res.body], {type: res.type});
+    if (typeof window.navigator.msSaveBlob !== 'undefined') {
+        // IE workaround for "HTML7007: One or more blob URLs were revoked by closing the blob for which they were created. These URLs will no longer resolve as the data backing the URL has been freed."
+        window.navigator.msSaveBlob(blob, filename);
+    } else {
+        const URL = window.URL || window.webkitURL;
+        const downloadUrl = URL.createObjectURL(blob);
+
+        if (filename) {
+            // use HTML5 a[download] attribute to specify filename
+            const a = document.createElement("a");
+            // safari doesn't support this yet
+            if (typeof a.download === 'undefined') {
+                window.open(downloadUrl, '_blank');
+            } else {
+                a.href = downloadUrl;
+                a.download = filename;
+                document.body.appendChild(a);
+                a.setAttribute("target","_blank");
+                a.click();
+            }
+        } else {
+            window.open(downloadUrl, '_blank');
+        }
+
+        setTimeout(function () { URL.revokeObjectURL(downloadUrl); }, 100); // cleanup
+    }
+}
